[NanaZip.Core] Update 7-Zip to 24.05. (Thanks to Igor Pavlov. Noticed by AVMLOVER-4885955 and PopuriAO29.)

+++ NanaZip.Core/SevenZip/CPP/7zip/Archive/VmdkHandler.cpp
@@ -815,7 +815,7 @@ Z7_COM7F_IMF(CHandler::GetArchiveProperty(PROPID propID, PROPVARIANT *value))
 
       bool zlib = false;
       bool marker = false;
-      int algo = -1;
+      Int32 algo = -1;
 
       FOR_VECTOR (i, _extents)
       {
@@ -829,7 +829,7 @@ Z7_COM7F_IMF(CHandler::GetArchiveProperty(PROPID propID, PROPVARIANT *value))
         {
           if (h.algo == 1)
             zlib = true;
-          else if (algo != (int)h.algo)
+          else if (algo != h.algo)
           {
             s.Add_Space_if_NotEmpty();
             s.Add_UInt32(h.algo);
