+++ NanaZip.Core/SevenZip/CPP/7zip/UI/Common/PropIDUtils.h
@@ -9,9 +9,9 @@
 void ConvertPropertyToShortString2(char *dest, const PROPVARIANT &propVariant, PROPID propID, int level = 0) throw();
 void ConvertPropertyToString2(UString &dest, const PROPVARIANT &propVariant, PROPID propID, int level = 0);
 
-bool ConvertNtReparseToString(const Byte *data, UInt32 size, UString &s);
-void ConvertNtSecureToString(const Byte *data, UInt32 size, AString &s);
-bool CheckNtSecure(const Byte *data, UInt32 size) throw();
+bool ConvertNtReparseToString(const Byte *data, size_t size, UString &s);
+void ConvertNtSecureToString(const Byte *data, size_t size, AString &s);
+bool CheckNtSecure(const Byte *data, size_t size) throw();
 
 void ConvertWinAttribToString(char *s, UInt32 wa) throw();
